requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.2.0
ftfy==6.1.3
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai-clip==1.0.1
packaging==23.2
pillow==10.2.0
pydantic==2.6.2
pydantic_core==2.16.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
timm==0.9.16
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
